anicet: add multiple runs
The current per-codec encoder operation includes the same process for
each codec:
* (a) codec setup
* (b) input conversion (to provide the right format to the encoder)
* (c) the actual image encoding (from raw blob to encoded blob)
* (d) codec clean up

For memory usage, we care about the 4 steps: If a codec requires allocating
100 MB during the codec setup, that needs to be accounted. For CPU usage,
we care only about the part (c).

Another part of the resource consumption tracking is that part (c) may
be too small to get interesting data. Moreover, at some point we want to
do actual battery measurements, and the total consumption of a single
image encoding may be too small. So what we want is to be able to do
(a), (b), and (d) once, but run (c) multiple times (e.g. 100 times)
and average the results.

This patch does the following:
* (1) add a new parameter (`num_runs`) to the anicet_run_<codec>()
  functions to specify the number of runs
* (2) separate from the anicet_run_<codec>() functions the exact
  part that does the image encoding (part (c) above). We will run
  the CPU usage around this function.

We need to be carefule with the video codecs (e.g. x265, svt-av1).
For them, we need to make sure we are doing all keyframes.