anicet: remove unneeded RGB conversion in jpegli
jpegli is converting YUV420p to RGB. This makes little sense,
as JPEG typically encodes the Y, U, V planes anyway. So we are
converting YUV to RGB, so that the jpegli API has to convert
RGB back to YUV.

jpegli has a raw data API, which allows setting YUV input.

Tested:
```
$ ./anicet --image /data/local/tmp/bin/img/lena.512x512.yuv420p.yuv --color-format yuv420p --width 512 --height 512 --codec jpegli --num-runs 2
```

And then checked the 2 images (they encode correctly).