anicet: rename the loaded x265-8bit-nonopt function
Rename the dlsym function from `x265_encoder_open` to `x265_encoder_open_215`.
Rationale is that that is the actual symbol name in the x265 shared library.

The root issue here is x265's API versioning mechanism to allow multiple
x265 versions to coexist in the same process.

When we tried to run x265-8bit-nonopt, it failed with:
```
x265-8bit-nonopt: Failed to load symbols: undefined symbol: x265_encoder_open
```

Checking the actual symbols in the library, we find:
```
$ llvm-nm -D libx265-8bit-nonopt.so | grep encoder_open
00000000000dadd8 T x265_encoder_open_215
```

The x265 library uses a versioned symbol naming scheme. In x265.h,
there's a macro that aliases x265_encoder_open to the versioned name:
```
 #define x265_encoder_open x265_encoder_glue2(x265_encoder_open_, X265_BUILD)
```

Where X265_BUILD is 215 in this build.

Why the difference?
* Optimized version (uses direct linking): The header file macros expand
  x265_encoder_open to x265_encoder_open_215 at compile time
* Non-optimized version (uses dlsym): We are dynamically loading the
  function, bypassing the header macros, so we must use the actual
  symbol name in the .so file: x265_encoder_open_215

Interestingly, when we checked all the symbols, only x265_encoder_open is
versioned. All the other functions (x265_param_alloc, x265_encoder_encode,
etc.) use unversioned names, which is why those dlsym calls did not need
to change.